fix: codereviews and refactor
