+++ must-gather/internal/api/types.go
@@ -8,14 +8,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
-// Logger interface for logging operations
-type Logger interface {
-	Log(format string, args ...interface{})
-	Begin(format string, args ...interface{}) func()
-	Warn(format string, args ...interface{})
-	Info(format string, args ...interface{})
-}
-
 // Config holds the configuration for must-gather collection
 type Config struct {
 	// DestDir is the root directory where all collected data will be stored
