+++ must-gather/internal/logstore/lokistack/gvr.go
@@ -1,0 +1,11 @@
+package lokistack
+
+import "k8s.io/apimachinery/pkg/runtime/schema"
+
+var (
+	lokiGVR = schema.GroupVersionResource{
+		Group:    "loki.grafana.com",
+		Version:  "v1",
+		Resource: "lokistacks",
+	}
+)
